Created separate package for AOP terminology question
